requirements.txt
autopep8==2.0.0
fastgrab==0.2.0
MouseInfo==0.1.3
numpy==1.23.5
opencv-python==4.6.0.66
PyAutoGUI==0.9.53
pycodestyle==2.10.0
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
python-libxdo==0.1.2a1
python3-xlib==0.15
pytweening==1.0.4
six==1.16.0
tomli==2.0.1
